requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
boto3==1.24.53
botocore==1.27.53
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coverage==6.4.3
cryptography==37.0.4
factory-boy==3.2.1
Faker==14.0.0
Flask==2.2.1
Flask-Auth==0.85
Flask-HTTPAuth==4.7.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
iso3166==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
packaging==21.3
password-strength==0.0.3.post2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycountry==22.3.5
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
schwifty==2022.7.1
six==1.16.0
SQLAlchemy==1.4.39
stripe==4.1.0
tomli==2.0.1
urllib3==1.26.11
Werkzeug==2.2.1
